Refactor sync and thoughtIndex to isolate side effects.
Rename thoughtIndex reducer to updateThoughts.
Move sync calls from other reducers to updateThoughts.
Move syncRemote to sync.js.
Convert autologin to localStorage instead of Settings.
