+++ src/initialSettings.html
@@ -55,20 +55,6 @@
         </ul>
       </li>
 
-      <li>Autologin
-        <ul>
-          <li>=readonly</li>
-          <li>=hidden</li>
-          <li>=options
-            <ul>
-              <li>On</li>
-              <li>Off</li>
-            </ul>
-          </li>
-          <li>Off</li>
-        </ul>
-      </li>
-
       <li>Tutorial
         <ul>
           <li>=readonly</li>
