+++ src/reducers/authenticate.js
@@ -1,8 +1,9 @@
 export default (state, { value, user, userRef }) => ({
-  // autologin is set to true in separate 'settings' action to set localStorage
+  // autologin must be stored in localStorage separately since it is not modified on every authentication
   // assume firebase is connected and return to connected state
   authenticated: value,
+  autologin: value || state.autologin,
   status: value ? 'loading' : 'disconnected',
   user,
-  userRef
+  userRef,
 })
