+++ src/reducers/settings.js
@@ -12,7 +12,7 @@ import {
 
 import existingThoughtChange from './existingThoughtChange'
 
-export default (state, { key, value, local, remote }) => {
+export default (state, { key, value }) => {
 
   const newValue = value.toString()
   const context = [EM_TOKEN, 'Settings'].concat(key)
@@ -33,7 +33,5 @@ export default (state, { key, value, local, remote }) => {
       value: newValue,
       rank: oldThoughtRanked.rank,
     }),
-    local,
-    remote,
   })
 }
